--- cmd/client/main.go
@@ -1,52 +1,0 @@
-package main
-
-import (
-	"bufio"
-	"fmt"
-	"io"
-	"net/http"
-	"os"
-	"strings"
-)
-
-func main() {
-	endpoint := "http://localhost:8080/"
-	// приглашение в консоли
-	fmt.Println("Введите длинный URL")
-	// открываем потоковое чтение из консоли
-	reader := bufio.NewReader(os.Stdin)
-	// читаем строку из консоли
-	long, err := reader.ReadString('\n')
-	if err != nil {
-		panic(err)
-	}
-	long = strings.TrimSuffix(long, "\n")
-	// добавляем HTTP-клиент
-	client := &http.Client{}
-
-	long = long[:len(long)-2]
-	// пишем запрос
-	// запрос методом POST должен, помимо заголовков, содержать тело
-	// тело должно быть источником потокового чтения io.Reader
-	request, err := http.NewRequest(http.MethodPost, endpoint, strings.NewReader(long))
-	if err != nil {
-		panic(err)
-	}
-	// в заголовках запроса указываем кодировку
-	request.Header.Add("Content-Type", "text/html; charset=utf8")
-	// отправляем запрос и получаем ответ
-	response, err := client.Do(request)
-	if err != nil {
-		panic(err)
-	}
-	// выводим код ответа
-	fmt.Println("Статус-код ", response.Status)
-	defer response.Body.Close()
-	// читаем поток из тела ответа
-	body, err := io.ReadAll(response.Body)
-	if err != nil {
-		panic(err)
-	}
-	// и печатаем его
-	fmt.Println(string(body))
-}
